requirements.txt
black==22.3.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
flake8==4.0.1
identify==2.4.12
idna==3.3
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.18.1
pycodestyle==2.8.0
pydotenv==0.0.7
pyflakes==2.4.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
